HW Keys customization support (2/2)
 *Based on DUI

jhenrique09 edits: make it DUI independent

Original commit message:
Same robust action library as used on software navigation. Supports
single tap, double tap, and long press. As a precautionary measure,
single tap back and single tap home are fixed and can not be changed.
Camera button actions are not supported at this time. We will bring
in wake key support at a later time.

Also includes:

Change-Id: I29ff38678821ca80db36d49d3b10d8ac29a6b4de
DUI: Initial checkin for Oreo [4/7]

Enable single tap rebinding for hwkey back and home

Initially, this policy was put in place to safeguard
the UX by preventing a usecase where the device could
be put in a state without back or home available. However,
if a user puts the device in this state, it is easy enough
to navigate to button settings to fix it

 *gzr: moved to androidx
       some cleanup

Change-Id: I6b71f9e54fcb99356e07f236e9ed926ac5fc017e